+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Only write entries that are worth mentioning to users.
 
 ## Unreleased
 
+- Shell: Fix image pasting from clipboard on macOS
+
 ## 0.86 (2026-01-24)
 
 - Build: Fix binary builds